+++ app/Http/Controllers/Admin/CategoryController.php
@@ -246,7 +246,7 @@ class CategoryController extends Controller {
                                 ->with('seo', 'products', 'blogs', 'freeWallpapers')
                                 ->first();
                 /* xóa ảnh đại diện trên google_clouds */ 
-                Upload::deleteWallpaper($info->seo->image);
+                if(!empty($info->seo->image)) Upload::deleteWallpaper($info->seo->image);
                 /* delete relation */
                 $info->products()->delete();
                 $info->blogs()->delete();
@@ -257,8 +257,8 @@ class CategoryController extends Controller {
                 /* delete các trang seos ngôn ngữ */
                 foreach($info->seos as $s){
                     /* xóa ảnh đại diện trên google_clouds */ 
-                    Upload::deleteWallpaper($s->infoSeo->image);
-                    foreach($s->infoSeo->contents as $c) $c->delete();
+                    if(!empty($s->infoSeo->image)) Upload::deleteWallpaper($s->infoSeo->image);
+                    if(!empty($s->infoSeo->contents)) foreach($s->infoSeo->contents as $c) $c->delete();
                     $s->infoSeo()->delete();
                     $s->delete();
                 }
