+++ app/Http/Controllers/Admin/PageController.php
@@ -215,14 +215,14 @@ class PageController extends Controller {
                                 ->with('seo')
                                 ->first();
                 /* xóa ảnh đại diện trên google_clouds */ 
-                Upload::deleteWallpaper($info->seo->image);
+                if(!empty($info->seo->image)) Upload::deleteWallpaper($info->seo->image);
                 /* delete relation */
                 $info->files()->delete();
                 /* delete các trang seos ngôn ngữ */
                 foreach($info->seos as $s){
                     /* xóa ảnh đại diện trên google_clouds */ 
-                    Upload::deleteWallpaper($s->infoSeo->image);
-                    foreach($s->infoSeo->contents as $c) $c->delete();
+                    if(!empty($s->infoSeo->image)) Upload::deleteWallpaper($s->infoSeo->image);
+                    if(!empty($s->infoSeo->contents)) foreach($s->infoSeo->contents as $c) $c->delete();
                     $s->infoSeo()->delete();
                     $s->delete();
                 }
